Extract subprocess runtime into dedicated microcrate
