+++ Cargo.toml
@@ -67,6 +67,7 @@ members = [
     "crates/perl-content-length-framing",
     "crates/perl-uri",
     "crates/perl-path-security",
+    "crates/perl-subprocess-runtime",
     "crates/perl-workspace-discovery",
     "crates/perl-corpus",
     "crates/perl-lsp",
@@ -103,6 +104,7 @@ allow = [
     "perl-ast",
     "perl-builtins",
     "perl-content-length-framing",
+    "perl-subprocess-runtime",
     "perl-keywords",
     "perl-lexer",
     "perl-regex",
@@ -238,6 +240,7 @@ perl-keywords = { path = "crates/perl-keywords", version = "0.10.0" }
 perl-source-file = { path = "crates/perl-source-file", version = "0.10.0" }
 perl-text-line = { path = "crates/perl-text-line", version = "0.10.0" }
 perl-content-length-framing = { path = "crates/perl-content-length-framing", version = "0.10.0" }
+perl-subprocess-runtime = { path = "crates/perl-subprocess-runtime", version = "0.10.0" }
 perl-lsp-protocol = { path = "crates/perl-lsp-protocol", version = "0.10.0" }
 perl-uri = { path = "crates/perl-uri", version = "0.10.0" }
 perl-path-security = { path = "crates/perl-path-security", version = "0.10.0" }
