+++ crates/perl-lsp-tooling/Cargo.toml
@@ -21,14 +21,15 @@ default = ["lsp-compat"]
 lsp-compat = ["dep:lsp-types"]
 
 [dependencies]
-perl-tdd-support = { workspace = true }
 perl-parser-core = { workspace = true }
+perl-subprocess-runtime = { workspace = true }
 moka = { version = "0.12", features = ["sync"] }
 lsp-types = { version = "0.97.0", optional = true }
 serde = { version = "1.0.228", features = ["derive"] }
 
 [dev-dependencies]
 perl-tdd-support = { workspace = true }
+perl-subprocess-runtime = { workspace = true, features = ["mock"] }
 criterion = "0.8"
 
 [[bench]]
